[Web] Synchronous server routing
