+++ core/datacap-server/src/main/schema/1.11.0/update.sql
@@ -3,3 +3,4 @@
 -- --------------------------------
 alter table `audit_plugin`
     add column `count` bigint(20) default 0;
+
