+++ core/datacap-web/console-fe/src/views/admin/history/AdminHistory.vue
