+++ core/datacap-web/console-fe/src/views/admin/pipeline/PipelineGenerate.ts
