+++ core/datacap-web/console-fe/src/views/admin/snippet/SnippetDetails.vue
