+++ core/datacap-web/console-fe/src/views/admin/snippet/SnippetGenerate.ts
