+++ core/datacap-web/console-fe/src/views/admin/source/SourceGenerate.ts
