+++ core/datacap-web/console-fe/src/views/admin/source/components/TablePreview.vue
