+++ core/datacap-web/console-fe/src/views/layout/common/components/LayoutHeader.vue
@@ -81,6 +81,7 @@ import Common from "@/common/Common";
 import {AuthResponse} from "@/model/AuthResponse";
 import router from "@/router";
 import config from '../../../../../package.json';
+import {createDefaultRouter, createRemoteRouter} from "@/router/default";
 
 export default defineComponent({
   name: "LayoutHeader",
@@ -95,6 +96,8 @@ export default defineComponent({
     const handlerLogout = () => {
       localStorage.removeItem(Common.token);
       localStorage.removeItem(Common.menu)
+      createDefaultRouter(router)
+      createRemoteRouter([], router)
       router.push('/auth/signin')
     }
     const language = 'zh_cn';
