+++ core/datacap-web/console-fe/src/views/system/function/GenerateFunction.ts
