+++ core/datacap-web/console-fe/src/views/system/menu/MenuDetails.vue
