+++ core/datacap-web/console-fe/src/views/system/menu/MenuGenerate.ts
