+++ core/datacap-web/console-fe/src/views/system/menu/components/MenuTree.vue
