+++ core/datacap-web/console-fe/src/views/system/role/RoleDetails.vue
