+++ core/datacap-web/console-fe/src/views/system/role/components/RoleAllocation.vue
