+++ core/datacap-web/console-fe/src/views/system/schedule/GenertateSchedule.ts
