+++ core/datacap-web/console-fe/src/views/system/template/GenerateTemplate.ts
