+++ core/datacap-web/console-fe/src/views/system/user/UserGenerate.ts
